+++ CHANGELOG.md
@@ -5,6 +5,7 @@ FEATURES:
 BUG FIXES:
 * postgresql: fix bug with applying cluster's settings in PostgreSQL 14
 * iam: return error when access bindings not applied
+* storage: fix bug of wrong handling permission error denied on reading policy
 
 ## 0.74.0 (April 19, 2022)
 ENHANCEMENTS: